feat: resolved many opinions
