+++ packages/module-workflow/src/server/actions/executions.ts
@@ -1,8 +1,6 @@
 import actions, { Context } from '@tachybase/actions';
 import { Op } from '@tachybase/database';
 
-import { updatedAt } from 'packages/plugin-mock-collections/src/server/field-interfaces';
-
 import { EXECUTION_STATUS, JOB_STATUS } from '../constants';
 
 export async function destroy(context: Context, next) {
