+++ packages/plugin-workflow-analysis/src/client/components/executionComponents.tsx
@@ -1,0 +1,30 @@
+import { useActionContext, useCollectionRecordData, useTranslation } from '@tachybase/client';
+import { getWorkflowDetailPath } from '@tachybase/module-workflow/client';
+
+import { Link } from 'react-router-dom';
+
+export const executionVersionColumn = () => {
+  const { workflowId } = useCollectionRecordData();
+  const { setVisible } = useActionContext();
+  return (
+    <Link to={getWorkflowDetailPath(workflowId)} onClick={() => setVisible(false)}>
+      {' '}
+      {`#${workflowId}`}
+    </Link>
+  );
+};
+
+export const ExecutionLink = () => {
+  const { t } = useTranslation();
+  const { id } = useCollectionRecordData();
+  const { setVisible } = useActionContext();
+  return (
+    <Link to={getWorkflowExecutionsPath(id)} onClick={() => setVisible(false)}>
+      {t('View')}
+    </Link>
+  );
+};
+
+function getWorkflowExecutionsPath(id: string | number) {
+  return `/_admin/business-components/workflow/${id}/executions`;
+}
