+++ packages/preset-tachybase/src/server/index.ts
@@ -82,6 +82,7 @@ export class PresetTachyBase extends Plugin {
     ['workflow-approval', '0.22.37', false],
     ['ai-chat', '0.23.8', false],
     ['department', '0.23.22', false],
+    ['workflow-analysis', '0.23.41', false],
   ];
 
   get localPlugins() {
